Fix default user language initialization to be null
- Updated `create_user.py` to remove the default language parameter when initializing a new user.
- Modified `init_user.sql` to set the `bot_language` field to `NULL` by default instead of "fr".

This change ensures that the language.get function prioritizes the user's language rather than the guild language. Previously, the default language was set to "fr", causing the language to be French for everyone. Setting the `bot_language` to `NULL` allows the correct language to be determined based on user settings.